+++ CHANGELOG.md
@@ -1,5 +1,45 @@
 
 
+## [0.6.2](https://github.com/Devil-Training-Camp/ldk-cli/compare/0.6.1...0.6.2) (2024-8-5)
+
+
+### 🐛 Bug Fixes | Bug 修复
+
+* **plugin-manager:** plugin manager choose pkg manager ([6ca45a0](https://github.com/Devil-Training-Camp/ldk-cli/commit/6ca45a024759059eec312f257c9eb91446031797))
+
+
+### 📝 Documentation | 文档
+
+* **docs:** docs ([5cdda01](https://github.com/Devil-Training-Camp/ldk-cli/commit/5cdda019ad4571157849722d11cc11d23d879ab0))
+
+# Changelog
+
+## 0.6.2
+
+### &nbsp;&nbsp;&nbsp;Bug Fixes
+
+- **plugin-manager**: Plugin manager choose pkg manager &nbsp;-&nbsp; by **grey_coat** [<samp>(6ca45)</samp>](https://github.com/Devil-Training-Camp/ldk-cli/commit/6ca45a0)
+
+##### &nbsp;&nbsp;&nbsp;&nbsp;[View changes on GitHub](https://github.com/Devil-Training-Camp/ldk-cli/compare/0.6.1...dev)
+
+## [0.6.1](https://github.com/Devil-Training-Camp/ldk-cli/compare/0.6.0...0.6.1) (2024-8-4)
+
+
+### 📝 Documentation | 文档
+
+* **docs:** docs ([0f033b8](https://github.com/Devil-Training-Camp/ldk-cli/commit/0f033b8f0cf0f59b27dc5b209e522206e4a815c2))
+* **docs:** docs ([c4dcd28](https://github.com/Devil-Training-Camp/ldk-cli/commit/c4dcd28dc5dd101703179c3b708524e6c93a1f42))
+* **docs:** docs ([6f0338c](https://github.com/Devil-Training-Camp/ldk-cli/commit/6f0338c081ca9c964a0231e687474564e7e1b95b))
+* **docs:** project doc ([2f5e713](https://github.com/Devil-Training-Camp/ldk-cli/commit/2f5e713f4f32db2ef99b679815f64aa86b46713e))
+
+
+
+## 0.6.1
+
+*No significant changes*
+
+##### &nbsp;&nbsp;&nbsp;&nbsp;[View changes on GitHub](https://github.com/Devil-Training-Camp/ldk-cli/compare/0.6.0...dev)
+
 ## [0.6.0](https://github.com/Devil-Training-Camp/ldk-cli/compare/0.5.0...0.6.0) (2024-8-2)
 
 
@@ -7,7 +47,7 @@
 
 * **create:** create default ([2cf6fe8](https://github.com/Devil-Training-Camp/ldk-cli/commit/2cf6fe87991e7690b1380b899405cfaecbde2218))
 
-# Changelog
+
 
 ## 0.6.0